End with parser

+++ src/model/figure.h
@@ -4,6 +4,10 @@
 #include<vector>
 #include<cstddef>
 
+
+//singleton
+
+
 namespace s21 {
 
 class Figure;
@@ -37,19 +41,19 @@ class Figure {
     void SwapVertex(std::vector<double>*);  //TODO is needed??
 
     size_t GetSurfacesCount();
-    const std::vector<double>& GetSurface(size_t n);
-    void AddSurface(const std::vector<double>& x);
-    void SwapSurface(size_t pos, const std::vector<double>& surface);
+    const std::vector<int>& GetSurface(size_t n);
+    void AddSurface(const std::vector<int>& x);
+    void SwapSurface(size_t pos, const std::vector<int>& surface);
 
   private:
     static Figure* self_;
     static FigureKiller killer_;
     std::vector<double>* vertex_;
-    std::vector<std::vector<double>>* surfaces_;
+    std::vector<std::vector<int>>* surfaces_;
 
     Figure() {
       vertex_ = new std::vector<double>();
-      surfaces_ = new std::vector<std::vector<double>>();
+      surfaces_ = new std::vector<std::vector<int>>();
     }
 
     ~Figure() {
